Adds JwtService and AclService to the API

+++ src/main/java/com/github/vaibhavsinha/kong/model/plugin/authentication/jwt/JwtCredential.java
@@ -6,6 +6,8 @@ import lombok.NoArgsConstructor;
 
 /**
  * Created by vaibhav on 16/06/17.
+ *
+ * Updated by dvilela on 17/10/17.
  */
 @Data
 @NoArgsConstructor
@@ -18,7 +20,7 @@ public class JwtCredential {
     @SerializedName("id")
     private String id;
     @SerializedName("created_at")
-    private Integer createdAt;
+    private Long createdAt;
     @SerializedName("key")
     private String key;
     @SerializedName("algorithm")
